# server/controllers/Beezvibes_db/generated/RoleControllerGenerated.js
/* 
* Generated by
* 
*      _____ _          __  __      _     _
*     / ____| |        / _|/ _|    | |   | |
*    | (___ | | ____ _| |_| |_ ___ | | __| | ___ _ __
*     \___ \| |/ / _` |  _|  _/ _ \| |/ _` |/ _ \ '__|
*     ____) |   < (_| | | | || (_) | | (_| |  __/ |
*    |_____/|_|\_\__,_|_| |_| \___/|_|\__,_|\___|_|
*
* The code generator that works in many programming languages
*
*			https://www.skaffolder.com
*
*
* You can generate the code from the command-line
*       https://npmjs.com/package/skaffolder-cli
*
*       npm install -g skaffodler-cli
*
*   *   *   *   *   *   *   *   *   *   *   *   *   *   *   *
*
* To remove this comment please upgrade your plan here: 
*      https://app.skaffolder.com/#!/upgrade
*
* Or get up to 70% discount sharing your unique link:
*       https://app.skaffolder.com/#!/register?friend=5dc416f1f1ef4518a537f625
*
* You will get 10% discount for each one of your friends
* 
*/
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE RoleControllerGenerated.js PLEASE EDIT ../RoleController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import RoleModel from "../../../models/Beezvibes_db/RoleModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import RoleController from "../RoleController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/role`;
    router.post(baseUrl + "", authorize([]), RoleController.create);
    router.delete(baseUrl + "/:id", authorize([]), RoleController.delete);
    router.get(baseUrl + "/:id", authorize([]), RoleController.get);
    router.get(baseUrl + "", authorize([]), RoleController.list);
    router.post(baseUrl + "/:id", authorize([]), RoleController.update);
  },


  // CRUD METHODS


  /**
  * RoleModel.create
  *   @description CRUD ACTION create
  *   @param Role obj Object to insert
  *
  */
  create: async (req, res) => {
    try {
      const result = await RoleModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * RoleModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id Role
  *
  */
  delete: async (req, res) => {
    try {
      const result = await RoleModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * RoleModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id Role
  *   @returns Role
  *
  */
  get: async (req, res) => {
    try {
      const result = await RoleModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * RoleModel.list
  *   @description CRUD ACTION list
  *   @returns ARRAY OF Role
  *
  */
  list: async (req, res) => {
    try {
      const result = await RoleModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * RoleModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id Role
  *   @returns Role
  *
  */
  update: async (req, res) => {
    try {
      const result = await RoleModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;