# server/swagger.yaml
...
openapi: 3.0.0
info:
  title: Beezvibes
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /idea/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Idea
      parameters:
        - name: id
          in: path
          description: Id Idea
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Idea
      parameters:
        - name: id
          in: path
          description: Id Idea
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
    post:
      summary: CRUD ACTION update
      tags:
        - Idea
      parameters:
        - name: id
          in: path
          description: Id Idea
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Idea"
  /idea/findByUserID/{key}:
    get:
      summary: CRUD ACTION findByUserID
      tags:
        - Idea
      parameters:
        - name: key
          in: path
          description: Id of the resource UserID to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Idea"
...
  /role/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Role
      parameters:
        - name: id
          in: path
          description: Id Role
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Role
      parameters:
        - name: id
          in: path
          description: Id Role
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
    post:
      summary: CRUD ACTION update
      tags:
        - Role
      parameters:
        - name: id
          in: path
          description: Id Role
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Role"
    
  #-----  User  -----
...
  /user/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
    
  #-----  User  -----
  /Users/{id}/changePassword:
    post:
      summary: Change password by admin API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordAdmin:
                  type: string
      responses:
        '200':
          description: OK
  /Users/{id}:
    post:
      summary: Update User API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    delete:
      summary: Delete User API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Idea:
      properties:
        _id:
          type: string
        IdeaDesc:
          type: string
        IdeaTitle:
          type: string
        IsDeleted:
          type: boolean
        UserID:
          type: integer
        likeCount:
          type: integer
        unlikeCount:
          type: integer
      required:
        -  _id
        - IdeaDesc
        - IdeaTitle
        - UserID
    Role:
      properties:
        _id:
          type: string
        IsDeleted:
          type: boolean
        RoleName:
          type: string
      required:
        -  _id
        - IsDeleted
        - RoleName
    User:
      properties:
        _id:
          type: string
        IsActive:
          type: boolean
        IsPaid:
          type: boolean
        RoleId:
          type: integer
        mail:
          type: string
        name:
          type: string
        password:
          type: string
        surname:
          type: string
        username:
          type: string
      required:
        -  _id
        - RoleId
        - mail
        - password
        - username